refactor: moved str_after function
